"Added product group selection to compare function"
This commit added the ability for users to select a product group when utilizing the compare function in the cmms.js file. This enhanced feature was implemented by querying product group data in cmms/views.py and adding an additional select HTML element to the Swal.fire function. The selected product group is now passed along with the originally used date in the preConfirm function. As a part of this update, the `productGroups()` function was added to get the data of product groups and the settings were modified for DB server connection while testing.