Add preview and excel options to Swal.fire component
This commit adds a new select dropdown to the Swal.fire component in cmms.js for users to choose between a preview or downloading an Excel file. A corresponding 'preview' field is also added to the return object of the preConfirm function. This allows the options to be handled accordingly on the server side. Users can now either receive a detailed HTML table on the application or download the requested information in Excel format directly.