Add compare function and view for stock taking
Extended compare functionality in the Computerized Maintenance Management System (CMMS) to include a new HTML template for displaying stock comparison data. Comment field was also added to the 'StockCountTrans' model for users to leave remarks. Unnecessary debug logs in 'cmms.js' were commented and upgraded the UI by displaying the comments and the stock take summary on the page directly. When the preview mode is selected, a verbose table with summary details is displayed, while in 'excel' mode, a file download link is provided. A 'countComment' function was also introduced to allow users to add comments directly on the UI. The database connection in 'settings.py' was switched to local for testing, and the PK and as_of parameters were added to the compare view in 'urls.py'. The changes will provide more insights into the stocks and improve user experience.