Refactor comparison function & comment compare link
Changed the name of the function triggered on stock comparison from 'cmms.compare' to 'cmms.compareTrigger' for better clarity. The direct link to the 'compare' page using a hardcoded date was commented out as it was unnecessary and could potentially lead to inaccurate results. It's now fully dependent on the 'compareTrigger' function.