Add tooltip variable for appropriate comment display
A tooltip variable has been introduced to correctly display comments related to different items while hovering over the bell icon. If the comment length is less than 0 or equals 'null' in the cmms.js, the tooltip will be set to an empty string, otherwise it'll show the respective comment. This change ensures correct information display and improves user interaction.