"Adjust layout and move export function in compare template"
The layout of the 'compare.html' template was altered, reducing the 'col-sm' size from 3 to 2 for Total Count, SYS Qty, QTY Diff, and Val Diff sections. This was done to create space for a new Tools and Details section. In addition, the "EXPORT" function, which was previously in the Total Count section, has been moved to this new section. These changes were made for better information organization and ease of use.