"Add issue field to stock module and enhance comment dialog
Added a new field, 'Issue', to the stock model in CMMS. This is to provide more context when entering a comment on a stock item. The related dialog in the frontend has also been updated, now requiring the user to select an issue type along with submitting a comment. This new feature aims to better classify the different issues that may occur with a stock item, thus making it easier to track, diagnose and handle these issues."