Add infographic modal to display data statistics and 'maximizeBrowserWindow' function
Added a modal window for displaying infographic data to the 'compare.html' template, enabling a detailed view of issue stats and total sales (if permitted). The modal includes two main blocks for issues stats displaying as a pie chart and total sales count. To facilitate a full-view user experience, a new JavaScript function 'maximizeBrowserWindow' was appended to 'cust.js', which can be triggered to enlarge the browser window to full screen.

This user-friendly data representation allows users to dive deeper into the statistical information and brings better readability and interaction to the system.