Update pie chart data and function enhancements
Updated the pie chart data in compare.html to be dynamic instead of static, enhancing the accurate representation of different count types: wrong_entries, sys_error, and lost. Simplified the data-bs-toggle button by replacing it with a simpler id named "infographic".

In views.py, modified how the counts are obtained to correspond with the updated pie chart data in the HTML. Now, the counts are fetched directly from the StockCountHD object rather than filtering by stock_count_hd_id.

Added two new functions in cust.js: exitFullscreen() for exiting the full-screen mode and added a trigger in trigger.js for the "infographic" click to open a modal, maximize the browser window, and exit the full-screen mode after 3 seconds.

These changes enhance the user interaction and data accuracy in the compare functionality.