Update modal layout and data metrics in compare.html
Modified modal layout to full screen for better user interface and expanded data metrics to include wrong entries, system errors, and losses. The charting data of entries and values also includes the newly added metrics. Added a downloadURI function in cust.js for the users to download the chart as an image. Updated the entries and values calculations in cmms to include not counted entries and its corresponding values. The cmms.js function now returns the header containing entries and values data.