Move tot variable declarations up in views.py
This commit moves up the total (tot) variable declarations in `views.py` from line 200-204 to line 185-188. This change is made to ensure the tot variables (tot_phy, tot_sys, tot_qdif, tot_vdif) are initialized before they are used in the 'excel' doc format. This addresses an exception raised when 'excel' doc is compared before the initialization of these variables.