Add file input for stock count
Added a feature that allows the user to upload a csv file for stock count, rather than entering them manually. Also, updated the HTML templates and JavaScript to handle this change, and updated corresponding URLs and view in Django. This feature will provide a quicker and more efficient way for users to load their stock count data into the system.