"Add getProduct() and update VIEW method in cmms class"
This commit adds a new function getProduct() in the cmms.js class to fetch product details from the database based on different attributes like barcode or item_ref. The default attribute is 'barcode'. The function makes an API call to '/cmms/api/' endpoint using 'VIEW' method.

In addition, the 'VIEW' method in the views.py file is updated to make a query based on the provided attribute (db_col) and its corresponding value. The attribute defaults to 'barcode'.