Update product info display and change database connection settings
Changes have been made to enhance the appearance and functionality of the product information display within the CMMS application. The code retrieving product details now also includes specific status messages and error handling. The addition of a table provides a more structured and readable layout for displaying product information. On the backend, we've replaced the term "barcode" with "item_uni" for data extraction which provides a more correct description. Spaces and unnecessary line breaks have been removed to improve the neatness of the code.

Additionally, the database settings in the configuration file have been adjusted to connect to a different database server. This change is made to match the current production environment.