Add last_breath field and update task
A new field called 'last_breath' has been added to the appscenter model. This field is updated with the current datetime when an AppAssignment object is created or retrieved in the apiv2 view. The reason for this change is to keep track of the datetime when an assignment was last active.