Add ref_freeze input field and update save button
This commit introduces a new input field 'ref_freeze' in the new_stock.html template, alongside adjusting the width of the 'remark' field to accommodate the change. Additionally, the Save button has been updated to invoke the 'cmms.saveFrozen()' function on click, enhancing the user interaction and functionality of the template.