"Renamed quantity field and enhanced item presentation in UI
Updated the 'quantity' field to 'qty' in cmms/models.py to make it consistent with referencing in other parts of the code. In cmms.js, input fields were reformatted to increase readability and ensure consistent data capture. The data presentation was also enhanced in new_stock.html for better visibility and user experience. Finally, an ownership field was added in views.py to associate each 'stock freez' entry with a user. This improvement facilitates better tracking and accountability."