Update form validation and data retrieval methods
Updated the form validation mechanism in the 'anton.js' file to enhance user experience. Alerts are replaced by border color changes in form fields--turn red on empty fields and green on filled ones. This is to provide visual feedback to users instead of disruptive alerts.

In 'cmms.js', corrected the property value used in getProduct() method from 'item_ref' to 'item_code' for accurate item retrieval. It resolves the incorrect item match issue which was based on 'item_ref' initially.