Increase width of 'NAME' field in PDF, remove 'FROZEN' field
In the PDF generation script, the width of the 'NAME' field has been increased from 80 to 100 to accommodate longer names. The 'FROZEN' field has been commented out as it is no longer needed in the report. This aims to improve the readability and relevance of the data in the generated PDF files.