Add 'NEW' option to dropdown menu
A new option has been added to the dropdown menu in the frozen.html template. This new option, labeled 'NEW', provides a quick and easy way for users to navigate to the 'new-frozen' URL. This was added to improve user experience by reducing the number of steps required to start a new task.