Adjust approval check in cmms.js
A condition was added to check if a transaction is not posted before it is approved. Previously, a transaction could be approved even if it was already posted. This could lead to discrepancies, as posted transactions should not be subject to further approval. This change ensures that only unposted transactions can be loaded for approval.