Add additional fields to views data and refactor PDF generation
In this update, we have appended the 'posted' status to the 'frozen' dictionary and the 'approved' status to the 'stock_hd' dictionary in 'cmms/views.py'. This provides better user information. Additionally, unnecessary code lines dedicated to pdf creation were eradicated to ensure cleaner code. The file name has been adjusted to correlate with the entry number rather than location ID. Lastly, the conditional statement has been updated for document type check 'STK' and an update to formatted response message was also implemented for better understanding and ease of debugging.