Add 'DELETE' method handling in views
In the views.py file, handling for the DELETE http method was added in. This was done to improve the functionality of the application by allowing users to delete specific stock records (either 'FR' or 'STK') denoted by a key. If the record exists, it gets deleted and a success message is returned. In case of an error during deletion, the error is returned. If a matching record can't be found, a 404 error is returned.