"Added stockExport function to cmms.js class"
Implemented a new function called stockExport in cmms.js file to enhance the stock module. This function calls the api with specific payload including the 'stage', 'doc', 'key' and 'document'. It further checks the status code of the response and presents a download file link to the user in case of successful api call (status code 200). This addition will provide users the ability to export the stock related data.