Add export options to count view card footer
Modified the card footer on the count_view.html to provide export options. The class list of the card footer div is changed to include 'd-flex' and 'flex-wrap' for better element positioning. Moreover, a new dropdown is added which allows the user to export the data as PDF, Excel, or CSV. This change provides the users more functionality while maintaining a clean UI.