Update API routing in url patterns
In the cmms/urls.py file, the api import path was changed to direct from the api folder, instead of the views folder. This change was made to ensure that API routes are properly handled by the api module. Also, code formatting was improved for better readability, especially in the '/sales' and '/customer' url patterns. This includes adding spaces after commas for consistency.