SALES CUSTOMER MODEL IMPLEMENTED 100Remove cmms_api.py to simplify structure
Removing cmms_api.py as the code in this file has been refactored and migrated to more specific file locations within our project structure. This action was taken to make our codebase more manageable and easy to navigate, reducing the complexity and improving maintainability. No functionality of the app is affected.