Adjust table layout and interactive elements for better usability
Updated the class list for the table in the 'count_edit.html' template to include 'datatable' for better compatibility with data libraries. Also reduced the width of input fields to increase visibility of table content and prevent overflow. Replaced button with icon for the 'line comments' functionality to enhance UI consistency and space efficiency. These changes aim to improve the user interaction and readability of the table.