Update counted_qty in cmms_api.
This commit corrects the issue where counted_qty could be less than 1 when exporting transactions. It introduces a conditional statement to check tran.counted_qty value. If the value is greater than 0, it is taken as counted_qty. Otherwise, counted_qty is set as 666. This ensures that the exported file always contains valid counted_qty values.