Add admin API for handling Geographical data
This commit introduces an API on the admin panel, which is specially designed to handle geographical data. This is achieved using Django and Python. The API allows us to perform CRUD operations on geographical data (GeoCity and GeoCitySub models), and returns a status report with each request. This feature aims to provide more structured and efficient management of geographical information on our system.