Expand geo module to support regions
In admin API file, support for geographical region was added allowing region to be handled the same as city. 'PK' look-up was changed from 'icontains' to 'equals' for accuracy while fetching records. Also, a unique identifier 'pk' was appended to the responses which will help with referencing the queried records. In views.py, the commented HttpResponse was uncommented to send form-based error responses which aids in debugging form-related issues.