Refactor DELETE method for added module functionality
This commit consists of refactoring the 'DELETE' method in 'cmms_api.py' to cater for different modules aside from the previous 'before' module. Previously, deletion was done only for documents 'FR' and 'STK'. This update includes a case where if the module is 'sales_customer', the code will delete the relevant sales customer. The 'cmms_cust.js' file was also updated to include a delete_customer function, which makes a DELETE API call to delete a certain customer. This allows for more flexible deletion depending on the module in question.