Update user interface and functionality in several templates
Several changes have been made in multiple templates.

In 'base.html', dynamic user profile pictures are implemented. Now, instead of the 'default.png', user profile picture is fetched from the server and displayed.

In 'emails.html', major improvements are done. For a better user experience, the email groups and email addresses sections, which are marks as 'col-sm-6', are hidden and only a table is shown which lists all the emails. This change should improve readability and declutter the interface.

Similarly in 'all-users.html', 'sign_up' url in the form action is replaced with 'create_account'. This change will route the form submission to the correct view thereby providing accurate results and eliminating any potential errors.