Add reminder feature and improve UI
Added a reminder feature in the admin panel which sends SMS reminders to users. These reminders check themselves based on their date and time and send an SMS to the user if they are due. A new entry path has been added in the urls for this.
Also, slight UI improvements have been made where the user profile picture is displayed. The 'alt' tags have been added for images which were missing. The image now appears round and with a shadow.
Various minor code refactoring were also conducted for better readability.