Add report functionality in meeting management module
New features include a user interface to generate meeting reports, backend interface to process and generate the desired reports. Changes also includes a new class for controlling the report card modal in Javascript. Adjustments were done on the meeting function to include the owner details in the returned JSON response.

This commit aims to improve the functionality of the meetings module by providing a way for users to generate detailed reports regarding their meetings. This feature allows them to better track their meetings, participants and the status of each meeting.