Add Reports and Taskmanager to installed apps, and update DB settings
This commit includes two new modules in the installed apps: 'Reports' and 'Taskmanager'. This is done to extend the functionality of our application.

Additionally, there were changes in the database settings. Previously commented out OLD_DB_SERVER and OLD_DB_PORT are now live with the actual values and new DB_SERVER and DB_PORT were added. This is tailored to fit the requirement of switching to a new database.