"Improve error handling & database fetching mechanism in cmms_api.py
This commit introduces two key changes in the cmms_api.py file. Firstly, it adds line numbers to the error messages, thereby making debugging easier in the future. Secondly, it adjusts the database query handling for customer and asset data, ensuring that we're using explicit cursor to execute and fetch data, increasing code readability and preventing potential issues with database connections. On the cmms_cust.js file, a console.table function call is commented out because it was not necessary and cluttered the console."