Add latest version download button in app page
A new feature has been added to 'app.html' that provides a button to download the latest version of an app. For the backend handling, a couple of methods ('latest_version' and 'latest_url') have been added to 'models.py' in 'appscenter'. This lets us fetch the latest version and URL to download the app. Also, an auto-generated config file for the taskmanager app has been created.