Temporarily comment out 'Meetings' and 'CMMS' sections in index.html
The "Meetings" and "CMMS" sections within the dropdown menus in our reports index.html template have been temporarily commented out. This is due to a reduction in their immediate relevance, as the focus is currently on implementing and testing new form designer functionalities. These sections will be restored once they are required or become relevant again.