"Remove IDE configuration files and optimize HTML code"
This commit deletes unnecessary IDE configuration files that were mistakenly tracked. These files are specific to user's local development environment and should not be part of the repository. Furthermore, a small change is made to the HTML code of the report index. Dropdown MenuButton IDs have been made unique for each legend to ensure proper dropdown functionality when there are multiple legends.