Update file upload paths in migrations
Updated file upload path in 'appscenter' migrations from 0084 to 0091. This change is necessary to be up-to-date with the file upload time and also to ensure that files uploaded do not overwrite each other. Additionally, added initial migrations for the 'reports' app for proper version control and replicate the application state across different environments.