Update filename in CSV export for STC document
The filename of the exported CSV in the case of a STC document now includes the reference number (ref) of the associated header. Previously, the filename only contained the location ID and remarks of the header. With the addition of the reference number, it will be easier to identify the corresponding header for each exported CSV file.