Add status field to SaleCustomerTransactions model and update related js and api
Added a new field 'finale' to the SalesCustomerTransactions model to track if a transaction was successful or failed. This required updates to the 'cmms_cust.js' to change the deal closure form and add the selected value to the payload. Additionally, the 'cmms_api.py' was also updated to retrieve and save the status of a deal in backend. This change helps to better evaluate the end results of each deal.